Kaung

+++ src/main/java/sg/nus/iss/team2/repository/LeaveRepository.java
@@ -3,6 +3,6 @@ package sg.nus.iss.team2.repository;
 import sg.nus.iss.team2.model.Leave;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaveRepository extends JpaRepository<Leave, Integer> {
+public interface LeaveRepository extends JpaRepository<Leave, Long> {
 
 }
