+++ src/main/java/sg/nus/iss/team2/service/UserService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface UserService {
     List<User> findAllUsers();
-    User findUser(Integer uid);
+    User findUser(Long uid);
     User createUser(User user);
     void removeUser(User user);
     User updateUser(User user);
